add module data util main
